Mostly finishes moving common functionality in bigm and hull to the base class, breaks hull tests because of transformed constraint naming
